+++ src/main/java/dev/shadowsoffire/apotheosis/data/TierAugmentProvider.java
@@ -9,6 +9,7 @@ import dev.shadowsoffire.apotheosis.tiers.augments.TierAugment;
 import dev.shadowsoffire.apotheosis.tiers.augments.TierAugment.Target;
 import dev.shadowsoffire.apotheosis.tiers.augments.TierAugmentRegistry;
 import dev.shadowsoffire.apothic_attributes.api.ALObjects;
+import dev.shadowsoffire.apothic_enchanting.Ench;
 import dev.shadowsoffire.placebo.json.RandomAttributeModifier;
 import dev.shadowsoffire.placebo.util.StepFunction;
 import dev.shadowsoffire.placebo.util.data.DynamicRegistryProvider;
@@ -33,13 +34,18 @@ public class TierAugmentProvider extends DynamicRegistryProvider<TierAugment> {
 
     @Override
     public void generate() {
+        // Haven
+        this.addAttribute("haven/max_eterna", WorldTier.HAVEN, Target.PLAYERS, 90, Ench.Attributes.MAX_ETERNA, Operation.ADD_VALUE, -70);
+
         // Frontier
+        this.addAttribute("frontier/max_eterna", WorldTier.FRONTIER, Target.PLAYERS, 90, Ench.Attributes.MAX_ETERNA, Operation.ADD_VALUE, -55);
         this.addAttribute("frontier/armor", WorldTier.FRONTIER, Target.MONSTERS, 100, Attributes.ARMOR, Operation.ADD_VALUE, 4);
 
         this.addAttribute("frontier/experience", WorldTier.FRONTIER, Target.PLAYERS, 100, ALObjects.Attributes.EXPERIENCE_GAINED, Operation.ADD_MULTIPLIED_TOTAL, 0.35F);
         this.addAttribute("frontier/luck", WorldTier.FRONTIER, Target.PLAYERS, 200, Attributes.LUCK, Operation.ADD_MULTIPLIED_TOTAL, 0.20F);
 
         // Ascent
+        this.addAttribute("ascent/max_eterna", WorldTier.ASCENT, Target.PLAYERS, 90, Ench.Attributes.MAX_ETERNA, Operation.ADD_VALUE, -40);
         this.addAttribute("ascent/armor", WorldTier.ASCENT, Target.MONSTERS, 100, Attributes.ARMOR, Operation.ADD_VALUE, 8);
         this.addAttribute("ascent/armor_pierce", WorldTier.ASCENT, Target.MONSTERS, 200, ALObjects.Attributes.ARMOR_PIERCE, Operation.ADD_VALUE, 5);
 
@@ -47,6 +53,7 @@ public class TierAugmentProvider extends DynamicRegistryProvider<TierAugment> {
         this.addAttribute("ascent/luck", WorldTier.ASCENT, Target.PLAYERS, 200, Attributes.LUCK, Operation.ADD_MULTIPLIED_TOTAL, 0.30F);
 
         // Summit
+        this.addAttribute("summit/max_eterna", WorldTier.SUMMIT, Target.PLAYERS, 90, Ench.Attributes.MAX_ETERNA, Operation.ADD_VALUE, -25);
         this.addAttribute("summit/armor", WorldTier.SUMMIT, Target.MONSTERS, 100, Attributes.ARMOR, Operation.ADD_VALUE, 12);
         this.addAttribute("summit/armor_toughness", WorldTier.SUMMIT, Target.MONSTERS, 150, Attributes.ARMOR_TOUGHNESS, Operation.ADD_VALUE, 10);
         this.addAttribute("summit/armor_pierce", WorldTier.SUMMIT, Target.MONSTERS, 200, ALObjects.Attributes.ARMOR_PIERCE, Operation.ADD_VALUE, 15F);