+++ src/main/java/dev/shadowsoffire/apotheosis/socket/gem/bonus/special/LeechBlockBonus.java
@@ -5,6 +5,7 @@ import java.util.Map;
 import com.mojang.serialization.Codec;
 import com.mojang.serialization.codecs.RecordCodecBuilder;
 
+import dev.shadowsoffire.apothic_attributes.api.AbilityCooldowns;
 import dev.shadowsoffire.apotheosis.Apoth.LootCategories;
 import dev.shadowsoffire.apotheosis.affix.Affix;
 import dev.shadowsoffire.apotheosis.socket.gem.GemClass;
@@ -36,11 +37,11 @@ public class LeechBlockBonus extends GemBonus {
     @Override
     public float onShieldBlock(GemInstance inst, LivingEntity entity, DamageSource source, float amount) {
         Data d = this.values.get(inst.purity());
-        if (amount <= 2 || Affix.isOnCooldown(makeUniqueId(inst), d.cooldown, entity)) {
+        if (amount <= 2 || AbilityCooldowns.isOnCooldown(entity, makeUniqueId(inst), d.cooldown)) {
             return amount;
         }
         entity.heal(amount * d.healFactor);
-        Affix.startCooldown(makeUniqueId(inst), entity);
+        AbilityCooldowns.startCooldown(entity, makeUniqueId(inst));
         return amount;
     }